refactor member search into separate function
